Use Godot Navigation Lite by Miloš Lukić (miloslu)
So the NavigationMesh was all well and good, BUT you have to bake the
Navmesh before the game starts. There is no tolerance for modifying
the navigation mesh on the fly in-game. For our procedurally generated
game, with city-building elements, that just wasn't going to fly.

Interestingly enough, dynamic and re-bakeable navigation meshes have
been implemented in Godot 4.0. But that's a "it's done when it's done"
project, and they're pretty firmly planted around 70%. That's probably
still 3 months out, at least. More like six months to a year, if I had
to guess. Hope I'm surprised, plan to be disappointed. Anyway, since
the Nightly builds are unstable, I don't want to risk damaging the
project. So while Godot 4.0 cooks, I had to look for another solution.

I saw a couple of alternatives, but I ultimately went with Godot
Navigation Lite by Miloš Lukić. It was really plug-and-play, and seems
to have everything I need. I swapped out the Navigation node we were
testing and replaced it with the DetourNavigation Node and client
meshes. It was almost seamless, save for a few caveats.

The DetourNavigationMesh doesn't offer a "get closest position"
function. That's pretty bad, given how much the Pawn and Driver code
rely on that adjusted/corrected position data. However, I realized
that getting the path provided us with adjusted points. So, I used a
cardboard-tape-and-gum solution: just call the path function FROM & TO
our current position. Voila: adjusted position!

I added a new configuration item for Pawn: the position_algorithm.
This variable controls how we calculate our adjusted/corrected
position. Before we were using a conditional, but I figured we should
go for some more explicit control.

In addition, you can now get the Pawn's algorithm adjusted position at
any time using the new get_adjusted_position() function. This is
basically just the old body text of a signal-response function moved
to a more accessible spot.