Build and change a NavigationMesh on the fly
This game/prototype is a city builder. It was a critical lynchpin,
then, that we be able to add terrain or delete terrain from the scene
and have the NavigationMesh update appropriately.

This is not possible in vanilla Godot 3.2.3. However, with the Godot
Navigation Lite plugin, it's possible to do EXACTLY this!

To test out this capability, I've added a new test scene:
DynamicNavMeshTest. This scene is almost a duplicate of
PawnPathingTest, but features controls for adding and removing
buildings on the fly (and a different building cycle). Whenever a
building is added or deleted, we rebake the DetourNavigationMesh.

We'll need to keep an eye on that as time goes on. If we end up with
a big map, there might be a big performance impact for adding a new
building. Or maybe there won't be? Hard to say...

Also, I tweaked House and Hut - rather than the "hiding corners" being
given in cardinals (i.e Northwest, Southeast), they are now given in
relative directional positions - i.e. Back Right, Front Left. I did
this since I was GENUINELY having a lot of trouble deducing what the
values actually meant (since I tend to use North-South as a reference
for camera movement).

I also reduced the Hut's cartoonishly tall roof. It was funny but it
was getting in the way of testing.