Set next phase to-do list
Trying to get the Pawn to path around, and giving it stuff to path
around, constituted Phase I of the prototype. That's really kinda an
arbitrary distinction, but I consider it a worthy one. After the
Boids, all of my efforts were were directed towards moving and
animating the Pawns. About six months later, we've got Pawns that are
animated, made by us, and can successfully navigate on-the-fly
bakeable navmeshes.

So, I thought it'd be good to lay out the next phase of development,
as outlined in our to-do list. First, I want to touch up some of our
existing structures. Then, do some test improvements, and finally work
on weapons and some way of handling attacks. That's the next major
hurdle to cross.