Add an option to "cap" the bottom of generated walls
When I made the WallBasic scene, I purposefully did not render the
bottom, only the top. Since the camera peers down at the world, this
mesh would never be seen. It would just be a waste of memory and
space.

However, we use the basic wall as an overhang to make a doorway in the
gap-wall. This leaves the underside of the basic wall exposed. And,
since there is no mesh on the bottom, this means light could pass
through the bottom of the wall, causing all sorts of lighting
problems.

The "Top" mesh of the basic wall has been changed to the "Cap" mesh,
and this terminology change has been propagated through the different
wall types. The basic wall now has a boolean configurable to enable or
disable generating the bottom cap.