Delete PawnPathingTest and the House
I felt like both the PawnPathingTest scene and the House scene were
taking up unecessary space.

The PawnPathingTest was there just to test how the Pawn interacted
with the Navigation Lite navmesh. With the DynamicNavMeshTest scene,
we were doing that AND some extra stuff. So why keep the older test
around? It made more sense to just delete it.

Maintaining both the House and the Hut was starting to feel like a
burden. They served the same purpose, but the House was just bigger
and more in the way. Not to mention that changes to common assets and
code required duplicate effort - and don't get me started on the fact
that the scripts were ACTUAL duplicates. It was just starting to get
annoying, so I removed it. I like the square shape of the hut, though
maybe we can change that roof texture...