Change cutaway wall appearance via columns
I was pretty happy with how our drop-down/cut-away walls looked. I
created a black material just for the cutaway-tops, and it looked
pretty good. There was just one element I wasn't happy with: the
columns.

I had designed the columns to be really simple - squares of a fixed
height with a fixed side texture. However, the bright side texture
was contrasting with the black cutaway texture, which broke the nice
illusion that the walls had simply be cut away.

Looking at it, I realized I had no choice but to allow for each face
of the column to be customized. I devised a new system: the column has
a primary material and an alternate material. Each face can be
designated as either a primary or alternate face, and will change it's
texture to match.

This is a bit of an odd solution, but we only needed two materials,
and assigning faces like this fit with that appropriately. In
addition, some of the faces need to be changed on the fly depending on
whether or not certain walls are present. Having this
primary-alternate scheme makes it easy to change things on the fly. No
swapping materials on the fly, just a simple change of enum values.

While working, I added light and shadow to the DynamicNavMeshTest. I
was trying to see if that would look better (it would save the
extensive modification of column). The test was insufficient, but I
kept the lighting changes anyway.

This completes my touching-up of how the houses look. I've updated the
to-do list appropriately. I've also added a new to-do item: a shader
to help us better see through walls and other such obstacles.