Add a TestMenu
As I've been adding tests, I've either been running them individually
(through the Play Scene button) or I've been manually adjusting the
project's settings so that my "current" test is the default scene.

This has started to get annoying, especially because I frequently mess
up the buttons and accidentally launch the wrong scene. I need to do
something different. My idea of different, in this case, is a root
menu for all of the tests.

The menu itself works great, but I have two concerns. Firstly: once we
are in a scene, there is no way for us to back out. I think we'll have
to make some sort of "shell scene" that comes complete with a pause
menu. Maybe.

Secondly, I added a loading screen to the test menu, but it the way
we're switching the scene right now blocks all threads. Apparently,
the solution to this is something called "Background Laoding". It's
good that there's an established solution, we'll see how easy it is to
implement though...