Add a loading screen
Previously, launching a test meant a quirky 5 second delay as the
scene loaded. That's well enough for debug, but I want to know that
the game is actually doing something! Besides, switching scenes is a
major part of this game. We should figure out how to test and load
these scenes now.

So, I've a made loading scene. It's meant to overlay whatever your
current scene is as the next scene loads. It provides all of the
mechanisms for loading the scene that you'll need.

However, I've noticed something weird. This doesn't FEEL noticeably
faster - it seems like most of the "lag" time comes from INSTANCING
the node, not loading the resources. If you think about it, that makes
sense: this game doesn't use models or large textures; it generates
what it needs on the fly and uses pixel art. It's that instancing, and
readying the scene, that is causing the hanging.

So our load screen shouldn't stop whenever the resource is done
loading - we'll have to look at how we're doing things and change it
up. For now though, this loading screen is a good start.