Add a pause menu and a testing shell
We could go from the root testing menu to a test, but not back to the
root testing menu. For that, I wanted to add a pause menu.

And add a pause menu I did! However, I realized that the processing
for pausing the game was pretty complicated. The tests have all been
designed to be very streamlined - quick to setup, quick to run, and
highly specialized. Having to recreate that code for each test would
really weigh down each test.

It would be better if there was a shell that could run each test and
handle the loading & pausing code appropriately. That idea morphed
into the TestShell.

I will say, while this code works for this one specific instance, I'm
pretty unhappy with it from an organization standpoint. Having to
manually manage multiple states (i.e. At Menu, Loading, Scene Running)
and even a substate (Scene Paused) creates a lot of messy code.

Other, future scenes - what we'll term "game-ready" - should have
their own logic for dealing with pauses and loading, instead of
relying on some parent scene to handle it. That's just bad form for
these nodes, and creates these weird interdependencies.

However, it's only 200 some-odd lines and we probably won't ever need
to touch it again. It helps the tests maintain their slim figure so
I call it fine for now and move on.