Use 'escape' to pause and un-pause
You press 'escape', the game pauses. You press 'escape', the game
resumes. This is a control pattern that has held for almost every PC
game for my entire life.

However, it only half worked in the last commit - you could open the
menu with escape, but you couldn't close it. That was because I had
escape bound to two separate input events - "ui_cancel" (which it is
by default) and a new value - "game_core_pause" (I wanted "pause" to
be explicit and separate).

I thought I was going to have to choose one and ditch the other, but
it turns out that the scene tree has a set_input_as_handled() that
allows us to stop an input event from ricocheting around unecessarily.
Fine by me!

Also, in an unrelated fix - I noticed that the loading circle was only
rotating on the first load. It seems like when you call stop_all() on
a particular Tween it clears all of the "registered tweens". Odd, but
a simple fix - we just need to declare the property tweens BEFORE
starting.