Update KinematicDriver via FuncRef, not signals
Back when we implemented the Pawns pathing via NavigationNodes, we ran
into a problem: paths were handed from the navigation node to the
pawn, and then from the pawn to the KinematicDriver. This presented us
with an issue: we had to get the Pawn's current position from the
navigation node and pass it down to the KinematicDriver, BUT it had to
be at the KinematicDriver's behest.

Not only is this kind of reaching upward difficult, it's just not good
programming practice. But I didn't see any other way of getting the
data we needed to the KinematicDriver. So I used signals (also not a
stellar choice) to prompt the pawn to update the driver's current
position.

However, while searching for a way to organize the GUI layers, I found
a help section about organizing scenes. I agreed with a lot of what it
said and, better yet, it offered me a solution for fixing the signal
nonsense that had been bothering me: FuncRefs, a feature of Godot I
had completely forgotten about.

The KinematicDriver has been changed - it no longer emits a
request_position signal. Instead, there is a new String configurable:
position_function_name. The idea is, the user provides the position
function's name as part of the KinematicDriver integration. I like
this because it works similar to selecting the DriveBody, which is
part of normal connection process for KinematicDriver.

These changes necessitated subsequent updates in the Pawn and Cubit
scenes.

Two issues remain - firstly, we have no way of telling the user when
the KinematicDriver has been correctly configured. I think it's pretty
simple, but others could find it complicated. So, I should add
warnings via _get_configuration_warning().

Secondly, having the user put in a String name for a function is
pretty awkward, I'm sure I can come up with a better way. I think it's
pretty complicated, BUT it's possible to generate a list of values for
Godot to choose from. And you can do it dynamically. The question is,
can you get a list of all functions defined in an object's attached
script? I'm not so sure...