Add configuration warnings to KinematicDriver
I don't think it's that bad, but I guess - in a sort of objective
sense - KinematicDriver has gotten pretty complicated. Again, I'd
still call it plug-and-play, but it can be pretty finicky: you need to
meet some specific requirements to have the KinematicDriver actually
work AND not crash.

So, I've added a _get_configuration_warning() function. This function
allows us to put a warning, in editor, when a tool script is
misconfigured in some way. This helps to document this particular
quirk of a quirky node, so now (hopefully) anyone who messes around
with KinematicDriver can get it set up with ease.

I did have to add two setter functions for the drive body and position
function, to ensure that the configuration warning updated whenever
the user made adjustments. Everything seems to be in order, though our
code is perhaps just a bit more messy.