Improve the warning message for KinematicDriver
I wanted to see if there was a way to make a dropdown list so that
users could choose the KinematicDriver's Position Function with a nice
list instead of manually entering a string (which is pretty janky).

The short answer is no.

The long answer is also no, but with extra steps. To do the above, we
need to:
    a) Get a list of all the methods for a mode
    b) Filter that to just methods defined by a script on a node
    c) Update the export statement with the new choices
Now, doing a) is easy, and b) would be hard but doable! Where things
fall apart is c): Godot just doesn't offer a way for us to make a
dynamic choice list. I believe this might change in Godot 4.0 (or
maybe it won't); for now, Godot only accepts explicit enumerative
declarations.

With that being the case, I figured I could at least improve the
warning message. We now explicitly request a Vector3-returning
function; I hope that proves sufficient...