Test a (very messy) xray/mask/peek-a-boo shader
When a Pawn walks behind a house, we've been dropping down the walls
so that the user can see the Pawn. I think this looks bad and
undermines the aesthetics of the houses. But, we still have to see
through the houses - the game has an isometric perspective, and hiding
characters from the player just because the perspective doesn't work
would make people VERY MAD.

I have two ideas to address this issue. The first option would be a
way to render what's behind the wall, through the wall, without using
our "collapsing" logic. I call this technique "Xray" or "Peek-a-boo"
but apparently the larger community calls this "masking", "culling",
or some combination of the two.

The second option would be just filling in only the Sprite with color
when it goes behind a wall. I thought that would be the easier of the
two BUT turns out the first one was easier!

So I created the XraySphere and a test to go with it. The XraySphere
has a material that takes a ViewportTexture and then renders that in
front of EVERYTHING. Give it a Viewport that shows an alternate,
wall-less world and... VOILA! YOU CAN SEE THROUGH WALLS!

I'm worried about the performance impacts, but I'm not seeing any
immediately so... hooray? It doesn't look perfect, especially when
the Pawn is in front of stuff - it's meant to be turned on and off as
visibility applies.

Originally I wanted to include the XraySphere into every Pawn by
default. I even coded up the Pawn taking in a viewport as a
configurable and setting up the bubble-texture via code. However, the
more I thought about it, the more I realized this put an unecessary
burden on the Pawn scene. It was forcing it to integrate into a very
specific kind of scene. We kind of already do this with the
Navigation configurable, but I feel like that's more general (and
optional) - the XrayBubble is VERY scene-and-setup specific.

We might add it later, we may make an inherited scene later on, or we
may glue the two scenes together at runtime (currently favoring the
last strategy). In recognition of all these possibilities, I modified
"Pawn" into "BasePawn".

Now, there's lot's to do - I think I need to clean up tests, clean up
the visual layers, modify some buildings, create that coloring
shader... the work goes on!