Add an alternate Xray shader, the Silhouette
The Xray Bubble is good, but draws a LOT of attention to itself. We
would most likely use it for player characters and other elements
worth highlighting very explicitly.

For stuff that we still need to keep track of, we need an alternative,
less all-revealing solution. That's where our new shader, the
Silhouette, comes in.

This shader takes in a viewport that should JUST be things to
silhouette. We then color in the silhouette-able object while
discarding the black around it.

Now technically speaking this really SHOULD NOT work. I have no depth
buffer and, ergo, the silhouette should be drawn in an effectively
random order on the screen. However, this isn't happening. I suspect
it might be a glitch, but the sprite is somehow always rendering on
top of the silhouette effect. Once the sprite passes behind an
obstacle, the sprite disappears and the silhouette renders. This might
also have to do with the sprites being set to Opaque Pre-Pass render
mode... maybe the sprites are just always rendering on top of the
silhouette? Maybe. Maybe...

The point is, this is still somewhat incomplete. It only works because
we're using it on sprites, which puts a kibosh on using a similar
technique on models until Godot 4.0 comes out. In recognition of that,
I've updated our to-do list. 4.0 really can't come soon enough...

Since we now have a bunch of different views that we have to render, I
tried to divy up the visual layers in a way I thought made sense.
They'll probably change in the future, so I won't put much effort into
documenting them for right now. Maybe it would be worth it to have some
docs for these more arcane choices, like the VisualLayer names,
PhysicsLayer names, the grounding theory behind the shaders.

Now, we do have a problem with the silhouetting - sometimes, the
silhouette peaks out from the sides of the sprite. If we try and
reduce the silhouettes, we begin to lose pixels inside the silhouette.
We're gonna need to change our sprites - posterize them, and lose the
feathering. I think we can do that in Blender...