Create clearer silhouettes using ImageMagick scripts
As I outlined in the last commit, I was having an issue with our
silhouettes. The feathered edges were making an exact outline
difficult. It was hard to find just the right value to cut away the
silhouette at without cutting into the body of the sprite.

I wanted to see if I could fix this in Blender. Maybe it's possible. I
gave up, because the workflow of Blender is anathema to my programmer
brain.

Instead, I decided to use an image-processing program, ImageMagick, to
handle the processing for my sprites. Since ImageMagick is a command
line utility, this allows us to write scripts for easy and repeatable
batch processing. So, I've created two scripts: a processing script
and a spritesheet script. The processing script drops the feathered
edges from the input sprites and replaces them with a solid black
outline. The spritesheet just packs the arguments into a spritesheet
(we haven't had dedicated utility for that, so this is nice).

Using our new scripts, I created a new pawn spritesheet. I made them
a neutral color because I thought it would go better with the shader,
though the more I think about it the less I think that actually
matters.

Now I'm interested to see if I can do more post-processing to get
better edge-highlighting on the sprites. I also want to see about
posterizing and possible dithering. I should also document the scripts
and what they do, including alternatives explaining the basic
techniques. And finally, I really want to change the silhouette shader
so that the alpha of our input color actually gets used.