Fix minor issues in ImageMagick scripts
Fixed some minor issues in the ImageMagick scripts.

I noticed that passing a blank string as a directory was passing the
directory test, so I added a new test to ensure that each argument is
NOT an empty string.

I also noticed that the outputs from the sprite processing script were
named incorrectly. Instead of output_%d.png, the file was coming out
as output%d_.png. This has been corrected.