Change the silhouette shader to use input alpha
I wanted to be able to control the opacity of the silhouettes directly
through the shader parameters. This would give us more flexibility
over always being 100% opaque.

I thought this was going to be more difficult, but it was actually
pretty easy. There's a whole section of conditionals that I missed
that made it real smooth. I'm getting better at these visual scripting
shaders!