Add visual layers to our building components
Our different shaders - specifically those that depend on different
viewports for through-the-wall-effects - rely heavily on interactions
between different visual layers. However, because the visual layers
are set at the mesh level, it was difficult to set the layers on our
building components. Worse still, these could only be set for
components globally, not for individual component instances.

The building components now have a "Render Layers 3D" configurable.
This configurable acts as a shell for the individual mesh visual
layers. They can be updated on the fly and changes will propagate
to building sub-components.

I also updated the Hut, but not the other buildings. Testing the
shaders out, it looks pretty good - I'm happy!