Add a new silhouette shader
I was pretty happy with the silhouette shader that we used to have,
where the whole silhouette was filled with color.

However, there was another version of the shader that I created while
developing the other shader. In that shader, the color was simply
multiplied over the existing sprite to create the silhouette, so that
it maintained all the pixel details. Now that I think about it, that's
not much of a silhouette but I still liked the look of it a lot. So, I
decided to create a separate shader just for this effect.

I've termed this new shader the "Silhouette Multiply" shader. To
differentiate it from the other shader, the other shader has been
renamed to the "Silhouette Fill" shader. Both shaders are now in a
dedicated "Special Effects" folder. This folder will probably end up
holding most of our shaders and other cool effects (like explosions).

I also deleted the old DepthShader. I had saved that as a WIP, but any
and all depth shader work is going to need to be postponed to
Godot 4.0. Once in 4.0 we can get depth from different viewports and
REALLY start doing some wacky stuff.