Use actual game-elements to test X-Ray Shaders
Previously, we were using a random amalgamation of cubes and the "Hut"
scene to test out the X-ray shaders. It occurred to me that, if we're
really going to use these X-ray shaders, then we're going to need to
test them on game elements - not just cubes.

To start, we've moved past just testing shader bubbles - we're testing
out all sorts of different shapes now. The test is really now just for
trying out X-ray shaders. So, I renamed it from "Pawn Shader Bubble
Test" to "X-Ray Shader Test". This better reflects the purpose of the
test.

Now, in the actual test itself, I've added a Farm scene and Woods
scene. I updated the Pine Tree to exist on the appropriate layer.
Since we still had "crushable" barley, and that is bad for testing out
silhouettes and other items, I created a quick-and-dirty "Strong
Barley" scene. Honestly the farm and woods scenes are gonna need their
own whole update but that's for further on in development.

I'm pretty happy with the result.