Add sprite override for BasePawn
Everytime we've created a new or alternate sprite sheet, we've had to
manually go down into the BasePawn scene and change the sprite sheet,
which changes it for every single scene.

I'm looking to have multiple kinds of sprites running around in the
next scene. Obviously going into the scene and changing things at that
level isn't going to work. And I don't want to have multiple scenes
for singular types like I did earlier on. My heart was in the right
place, but those scenes started to differentiate and got out of hand
really quickly.

Instead, I'm just going to change what I need, and only what I need:
the sprite. To accomplish this, I've added a SpriteFrames override
configurable to the BasePawn. This overrides the sprite frames at
readytime, and also saves the original frames for later. It provides a
setter method for easy on-the-fly changes. Overall, I'm pretty happy
with it.