Test how X-Ray shaders interact with each other
We've been manually switching between our different shaders to observe
their effects. However, we've been unable to observe how shaders
behave when interacting with each other in a near-game facsimile.

So, I added a whole bunch of Pawns. I modified them to use a variety
of sprites (to help tell them apart) and attached different shaders.
This included two multiply shaders, two fill shaders, and an X-ray
shader. The shaders were modified with various colors and alphas to
help further differentiate the Pawns.

To order the pawns around, I added a "Telepointer" and new processing.
Whenever a pawn completes a path and asks for a new one, the
telepointer moves randomly to assign a new path to the requesting
pawn.

Deciphering exactly what was happening was difficult because all the
pawns moved so quickly. While that may be a problem in future builds,
it's just a minor annoyance here. So I added a GUI slider to allow for
fine control of the Engine speed. Easy slow motion.

Running the actual test, here's what I observed - any two shader
shapes were generally fine overlapping with each other out in the
open. Whenever two shaders overlapped behind a building, whatever
shader was considered to be in "front" rendered in front. That sounds
obvious but considering that we purposefully disabled depth testing
for the shader materials, I wasn't sure exactly what would happen.

This has important implications with how the X-Ray shader interacts
with the Silhouette shader. The X-Ray shader is meant to be a clear
view through walls. However, if a Silhouette shader happens to be in
front of an X-Ray shader, then the silhouette fills in over the
appropriate spot in the X-Ray shader. This is particularly noticeable
with the Fill shader, and is apparent, but less noticeable, with the
Multiply shader. We'll need to see if it's possible to force the X-Ray
to render in front.

Also interesting - Silhouettes shaders will shade anything behind
them. We just need to reduce the size of the material shader. All that
being said, there's just a lot of problems we could solve with
Godot 4.0, but will need work arounds for now. Who knows when that
will happen...