Add a dedicated camera rig scene
Cameras were, previously, implemented per-scene. That was mostly fine
since the code was simple, copy-paste kind of stuff. Plus, we've only
really been doing tests, so there was no real impetus to make
anything better or more advanced.

However, with the X-Ray and Silhouette shaders, we now have two
viewports and child cameras that need to be kept in sync with the main
cameras. Worse still, we'll need to duplicate this set up, and the
code, to any test that needs access to the X-Ray or Silhouette
shaders.

It only makes sense, then, to have a dedicated "Camera Rig" - which
this commit adds. The Camera Rig handles synchronization between the
main camera and the various viewport cameras. It offers built-in
movement and zoom handling, as well as various configurables for
movement and zoom. There's even optional position clamping on X and Z,
so we can limit where the camera goes.

I'm very happy with the output. By enabling the "Editable Children"
option in-editor, we can gain east access to the viewports and the
whole thing is pretty much seamless. I'm really happy with it!