Add a mesh to hide/show itself based on occlusion
As I (think) I outlined in previous commits, the X-Ray shader being
"always on" presents a problem: we only want it rendered when it's
behind walls. However, without access to multiple viewport depths,
it's impossible for us to do this with a shader. Our only option,
then, was to do it manually.

First, we needed to designate a collision layer just for items that
could obscure the camera. This new layer is the "Camera Obstruction"
layer. This needs its own layer; it would be inappropriate of us to
automatically assume any other collision layer will automatically
correspond to object occlusion. After all, the "shadow only" mode
already actively messes with that.

Speaking of, the building components that supported "shadow only"
mode were also updated. They now toggle collision on the "Camera
Obstruction" layer, which is enabled by default.

To make this nonsense easier on ourselves, I created a dedicated scene
to act as the X-Ray/Silhouette mesh which performs it's own camera
detection raycasts. It comes with an inversion option (show only when
in sight), a NECESSARY node path, a collision configuration option (if
we end up needing a mix of physics layers), and an animation length.
If the animation length is negative or zero, then the mesh simply
toggles visibility. Otherwise, it grows and shrinks the mesh over the
specified time interval. I think it's neat.

Nice thing about this scene - since the mesh is the root node, it
allows us to change the mesh in whatever the parent scene is.

So using physics layers to make strictly visual decisions irritates
me. A lot. It's the exact kind of mixing of different types and layers
that can lead to confusion down the road. Unfortunately for me and my
code composition complex, this solution works. Like, really well. I'm
not even sure we'll ever end up switching to the depth shaders - I was
too worried about the performance impact about multiple Viewports, and
I feel like everything is working great right now.

While I feel like the performance impact of constant raycasts is
something worth worrying about, I feel like this is a more platform
friendly solution.