Better control the Hiding Mesh's state
I added some more processing to the HidingMesh, in an attempt to make
the current state more obvious and provide more fine control over that
state

But that's not the real story behind this commit. The only reason I
made these changes was because I was messing around with the two
different kinds of pine trees: the normal, which uses AnimatedSprites,
and the simple, which uses meshes. I've always preferred the sprites
and I realized that, with one of the latest Godot updates (either
3.2.3 or 3.2.2), the major performance drawback of sprites has been
eliminated. Why not, then, switch back to using the normal trees?

Of course, it couldn't have been that easy. Because the trees were
sprites, they were rendering in FRONT of the X-Ray bubble. GAR! This
update to the HidingMesh only happened because I couldn't figure that
out.

I am dedicated to figuring out how to do this though. We could use a
depth shader in Godot 4.0 but that's too far away and I am too worried
about the performance. This is going to require us to reconsider how
we think about the X-Ray shaders.

You know, since the game is isometric, actual distance to the camera
doesn't matter for size - everything just has to be on the same line
for it to work. By that logic, we don't REALLY need to have the X-Ray
shader-shapes sat on the Pawns themselves, do we? No... we could just
have them float above it all...