Experiment with viewport shaders
As outlined in the last commit, I was having problems with how our
"X-Ray" shader was interacting with transparent/alpha sprites.

I came up with three different solutions - and three different shaders
to match. I must address something first though: we've used various
names for the X-Ray and Silhouette shaders, and have been somewhat
inconsistent about the terms. From now on, we'll be referring to these
shaders (and any other viewport sampling shader) as Viewport Shaders.

Since Viewport Shaders need to be local to the scene - i.e. they must
be defined in the scene (it's a complication of viewport textures) -
the shaders we've created aren't really actualized shaders as much as
they are templates. So I've collected everything in a Viewport Shader
Template directory.

Anyway, my first solution was to enable depth testing and move
whatever mesh we were working with into a camera-oriented line far
above wherever the Pawn is. That worked excellently! This new shader
is just called XrayShader - the old one is now called
XrayNoDepthShader.

I did run into a problem - since the HidingMesh was only casting from
it's origin, moving it off the Pawn was causing it to be always
hidden. Since it was far above where any obstacles were, there was no
collisions, and the mesh wasn't showing itself. To fix that, the
HidingMesh now has the option to cast from a dedicated Spatial.

I then had a new idea for a shader - instead of just using a mesh for
X-Ray shaders, what if we could use the alpha from a texture? This
would allow us to do more interesting stuff, like dithered visibility
textures. So I implemented that as XrayAlphaTexture and added a whole
sweeping of alpha masks. I made the alpha masks via a combination of
gradient and brushes in GIMP. I then fed these into the ImageMagick
poster-dither scripts.

Unfortunately, I was getting problems where the sprite trees were
rendering in front of meshes with the Xray Alpha shader attached. And,
thinking about it, that kind of made sense. All of our problems have
really bloomed from the alpha-depth-sorting-thing, so it made sense
that putting alpha on the mesh caused this problem to appear.

Turns out, a lot of this work wasn't necessary - because all I had to
do was force the shader to render last by increasing the
RenderPriority to 1. I did that for the XrayNoDepthShader and it works
now so... oh well! I like all of what we got, so I'm pretty happy.

I updated the XrayShaderTest - which is now called ViewportShaderTest
- so that each of our five viewport shaders are represented. Neat!