Use Sprites for pine trees, not Meshes
As I outlined in previous commits, I really wanted to get rid of the
two near-duplicate versions of PineTree that we had.

I had originally created a 'Simple' version because the Sprites in
Godot were implemented poorly and presented a pretty severe
performance impact. The 'Simple' version used a mesh to avoid this.

However, Godot 3.2.3 eliminated this burden. That meant we could
switch back to the Sprite version. I wanted to use the Sprite version
because it allowed for animations - perhaps we could add wind blowing
leaf animations or night transition animations. I also prefer the
Sprite's manual "pixel size" control.

So, I deleted the PineTreeSimple scene and it's assets.

I also updated the original scene with a new sprite that acts as the
Obstacle Alternate sprite - i.e. what SHOULD be shown through the lens
of one of our "X-Ray" shaders.

And with that, I think I'm satisified with the Viewport Shaders! I've
updated our to-do list. I want to take a second and breathe, do a
documentation push so that the weird intricacies of this project don't
get lost.