Add documentation for the ImageMagick scripts
As I outlined in the to-do list, I wanted to add documentation to the
ImageMagick scripts. It was undocumented, and the docs would help
other people work with the scripts and also help ME remember how the
scripts work.

So I've added documentation, including some images that have been set
aside for examples. One interesting choice I made was to write the
documentation for someone who was unfamiliar with bash scripting. I
did this because, if this project becomes the learning tool that I
want it to be, then assets will probably be the first thing people
modify. That's where I started, so long ago - making custom assets.
Poorly.

So it was important for me that the documentation should acquaint
people with what they're looking at and, better yet, describe the
steps they can take to reproduce the same result using software OTHER
than ImageMagick.