Add an OpenSCAD README
Added an OpenSCAD README to to document several things:
- The place of the scripts in our art/asset pipeline
- The reasoning behind using OpenSCAD
- OpenSCAD's unusual syntax

I WANTED to add a section about what the individual scripts do, but
there's only one script, it's currently unused, I dislike it, and
I intend to delete it at some point. So not worth the effort - we'll
expand the README with that section once we start adding scripts
(which will be part of the weapon-attacking effort).

This concludes the scripting documentation push - now we need to
start documenting the proper Godot project itself. Though I'm sure a
more elegant solution exists, I'm going to just document-in-place via
README files.